fix: 物体の色を計算するロジックをmain関数から分割
